pyproject.toml
[project]
name = "autonomous-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv==2.1.3",
    "beautifulsoup4==4.12.3",
    "chromadb==0.5.5",
    "duckduckgo-search==6.3.4",
    "langchain-core==0.3.19",
    "langgraph==0.2.35",
    "lxml==5.3.0",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "reportlab==4.2.2",
    "requests==2.32.3",
    "streamlit==1.37.1",
    "tavily-python>=0.7.21",
]
